+++ Runtime/Vector4Set.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+namespace UnderLogic.Collections
+{
+    [CreateAssetMenu(menuName = "Collections/Sets/Vector4 Set")]
+    public class Vector4Set : RuntimeSet<Vector4> { }
+}